requirements.txt
paho-mqtt
FUTIL
gTTS
pyttsx3
SpeechRecognition
PyAudio
nltk
retinasdk
